Don't skip over the first Tera error
The first line in the Tera error messages is sometimes less useful, but
in other cases, like when the context is not a map or struct, contains
the complete description.

Always include it, even if the output is slightly uglier. Also don't
append periods at the end, since some Tera messages already have them.