Improve Debug impl for Config by using `debug_struct()`
Using the method `debug_struct()` of `fmt::Formatter` takes
care of the exact formatting for us. Additionally, it also handles
the "alternate" form of printing enabled with '#'. In the struct
case it prints the struct on multiple lines instead of just one.

Note that this commit changes the output slightly. Before, the
field `log_level` was printed with `log: {}`. I assumed this
is a simple oversight and replaced "log" with "log_level".
Additionally, the value of `environment` is now shown as a simple
struct field instead of being combined with the struct name.