Fix errors![...] macro
In most applications, Rocket "lives" in the parent module. But sometimes, it
might be useful to encapsulate all Rocket code in a distict module like
`my_crate::http`.

In that case, rocket needs to be addressed with a absolute path instead of a
relative path. While the routes![...] macro does this correct, the errors![...]
macro did not.

This commit fixes the errors![...] macro.