fix: Make form items parsing less strict
Currently Rocket is pretty strict on form items parsing, which affects how parameters without a value in query strings are handled, for example for `?flag` or `?f&foo=bar`. The [RFC 3986](https://tools.ietf.org/html/rfc3986#page-23) do not impose any constraints to the usage of empty query parameters and it seems like the way most framewors and servers handle this is by assigning an empty value to the parameter ([ref](https://stackoverflow.com/questions/4557387/is-a-url-query-parameter-valid-if-it-has-no-value)).

This change modifies the `FormItems` iterator to handle empty or unvalued parameters.

This affects the limits handling as well, since a `POST` request with a form content type and a body `valu`, as it results from limiting `value=Hello+world` to 4 bytes, will produce a valid form items collection like `&[("valu", "")]`. Then when trying to parse it agains a given type without a `valu` field this will result in a `422 Unprocessable Entity` error instead of a `400 Bad Request`. See the change to the `limits_tests::much_too_small` test.