feat(codegen&core): Custom Requests
Using a string literal for the first part of the request decorator will mean that it is interpreted as a custom method. Rewrote tests to test this as well as properly accommodate it. Added example for custom methods.

Method.as_str() now returns a str with the lifetime of the method, not static. StaticRouteInfo now stores the method as a string & it is converted to/from when appropriate. Method::from_str() will never fail, however still returns a Result.

Fixes #232