Allow base16-encoded secret keys
Why
---

It's basically a workaround for Heroku, which has built-in support for generating secret keys,
but generates 256-bit hex-encoded strings, instead of 256-bit base64-encoded ones.

Right now, I have to work around it in my Procfile, which isn't bad, but it adds embarrasing friction
to what would otherwise be a pretty clean 12-factor app wrapper.
<https://github.com/notriddle/more-interesting/blob/ac944cc97c26571f799d090bb8dae78c9a3b98cb/Procfile>

What alternatives are there?
----------------------------

We could, instead, lift the format requirements entirely. We'd run PKDIF2 or something on the
secret_key, generating a 256-bit secret no matter what the input actually is.
I don't think that's better, because it means the user could use predictable, memorable "passwords"
instead of something truly random.

Leaving this alone is also an option, but it means punting to the application to handle converting
this, and getting it wrong can result in subtle disasters (like generating a secret of `0` because of
a typo).