Remove use of `!` type in favor of `Infallible`
For the most part, this is simple substitution and nothing else. In one
specific location (core/lib/src/config/mod.rs), the compiler threw an
error, as the `Infallible` type is not a special case in the compiler.
As such, we have to explicitly tell it that code after that point cannot
be reached using `unreachable_unchecked()`.