Migrate over to AsyncSeek
This is a fairly complicated change, because the rest of the framework
wants to know the size of the sized body right away, but we need to
fetch it asynchronously. I'm not sure if this is an elegant-enough
solution.

This commit changes the `Response::set_sized_body` in the obvious way,
but to preserve the builder pattern, `ResponseBuilder` is changed so
that the act of finalizing it is asynchronous, rather than attempting to
make the `sized_body()` function asynchronous.