impl (de)serialize for UUID
closes #1370

TODO: Think some discussion for the other options in the issue should be
discussed there before merging. This PR is mostly there for checking
what the implimitation could look like.

Also think we hit a limitation in cargo where we cant describe that we
want the feature rocket_conrtib::serde to be connected to the feature
flag of uuid::serde.