codegen: fix clippy::no-effect-underscore-bindings
This lint checks for binding to underscore prefixed variables
without side-effects.

Further usage of such variables is not checked, which can lead
to false positives if they are used later in the code.

To fix the false positive we use Rust's `r#` variable prefix
to avoid collisions with Rust keywords.