Correctly flush when persisting temporary files.
Operations on `TempFile` that persisted buffered contents to disk used
tokio's `File` struct to write the contents to disk. Unfortunately,
dropping a `File` at the end of the scope does not wait for the writes
to complete, since this could block.

When reading from the file immediately, it is possible for the contents
to be missing and for the file to appear empty. This is actually more
likely than not if the read is done synchronously.

This can easily be reproduced with the following snippet:
```rust
let mut f = TempFile::Buffered { content: b"Hello" };
f.persist_to("file.txt").await.unwrap();
assert_eq!(std::fs::read("file.txt").unwrap(), b"Hello");
```

This is behaviour is mentioned in [tokio's documentation](https://docs.rs/tokio/latest/tokio/fs/struct.File.html):
> A file will not be closed immediately when it goes out of scope if
> there are any IO operations that have not yet completed. To ensure that
> a file is closed immediately when it is dropped, you should call `flush`
> before dropping it

An awaited call to flush would solve the issue. However the fs::write
function performs the same sequence of operations in fewer lines of
code, and removes this particular footgun.